add base Card type
Together with the Card type, we're also creating the CardColor type.
As of now, we only have the 'face' and 'color' properties of the card.
The CardColor type is an enumeration, meaning it only has values
associated with names, and no behaviour at all. The color values are
defined as:
RED:    1 (0001)
GREEN:  2 (0010)
BLUE:   4 (0100)
YELLOW: 8 (1000)
ANY:   15 (1111)
so that one can use bitwise 'and' operations to check whether a card
can be placed on top of another one. Only change-color and +4 cards
have the color ANY.