implement Card.can_be_placed_onto(card)
This is the method that tells us whether we can place a card on top
of another card. If at least one of face or color is equal to the
one of the other card, then it can be stacked.

Also, changed the definition of CardColor.ANY from 15 to
RED | GREEN | BLUE | YELLOW. Read as: ANY can be any out of these
four.