implement Player.choose_random_valid_card
To make a CPU player choose a card, we first have to get
all playable cards available in its hand. We build a list of
valid cards, checking each card against the card on top of
the stack of played cards. If the player has at least one
valid card, then randomly choose one of them, remove it
from its hand, and return it. Otherwise, return None.

In this commit we also fix a bug previously introduced in
Card.can_be_placed_onto: a card can be placed on top
of another one as long as they have the same face OR
the same color, not both.