implement Player.play_best_valid_card
We start by defining Player.__get_playable_cards and moving the list
comprehension from Player.play_random_valid_card[1] to it. Since both
methods need to know them, it makes sense to have a method do just
that.

The logic behind choosing the best available card, is to choose the one
that is in excess in the player's hand: if the last played card is
yellow, and the player has many yellows, then it's better to play a
yellow card instead of a card of same face, but different color in order
to keep a variety of cards. Likewise, if the card is a 4, and the player
has many 4s, then it's better to play one of them instead of a
same-color card. The algorithm then proceeds to choose the best
combination of face/color to play.

Having a list of valid cards, we begin by creating two lists from the
original one: one sorted by color and face, and one sorted by face and
color. We group them by their respective orders to separate each pair of
color/face in their own lists. The list sorted initially by color only
has cards of the same color of the card on top of the stack, while the
list sorted by face first only has cards of the same face of that card.

Then we choose the one list with more elements, and finally, we choose the most
frequent face (if sorted by color first) or color (if sorted by face
first), remove it from the player's hand, and return it.

[1]: Renamed from choose_random_valid_card